requirements.txt

evalutils==0.3.1

# scikit-learn==0.24.2
# scipy==1.7.3
monai
pytorch-lightning
nibabel
# easydict

